+++ R/util_plot.R
@@ -1,4 +1,4 @@
-#' util_plot
+#' show_landscape
 #'
 #' Plot a Raster* object with the NLMR default theme (as ggplot).
 #'
@@ -14,38 +14,42 @@
 #' @examples
 #' \dontrun{
 #' x <- grdmap
+#'
 #' # classify
-#' y <- c(0.5, 0.15, 0.25)
-#' y <- util_classify(x, y, c("1", "2", "3"))
+#' y <- util_classify(grdmap,  n = 3, level_names = c("Land Use 1", "Land Use 2", "Land Use 3"))
 #'
-#' util_plot(x)
-#' util_plot(y, discrete = TRUE)
+#' show_landscape(x)
+#' show_landscape(y, discrete = TRUE)
 #'
-#' util_plot_grey(x)
-#' util_plot_grey(y, discrete = TRUE)
+#' show_landscape(list(grdmap, rndmap))
 #' }
 #'
-#' @aliases util_plot
-#' @rdname util_plot
-#' @name util_plot
+#' @aliases show_landscape
+#' @rdname show_landscape
+#' @name show_landscape
 #'
-NULL
+show_landscape <- function(x,
+                           xlab,
+                           ylab,
+                           discrete,
+                           div_scales,
+                           n_col,
+                           n_row,
+                           ...) UseMethod("show_landscape")
 
-#' @rdname util_plot
+#' @name show_landscape
 #' @export
-util_plot <- function(x,
-                      xlab = "Easting",
-                      ylab = "Northing",
-                      discrete = FALSE,
-                      ...) {
+show_landscape.RasterLayer <- function(x,
+                                       xlab = "Easting",
+                                       ylab = "Northing",
+                                       discrete = FALSE,
+                                       ...) {
   # derive ratio for plot, cells should be a square and axis equal in length
   if (raster::ncol(x) == raster::nrow(x)) {
     ratio <- 1
   } else {
     ratio <- raster::nrow(x) / raster::ncol(x)
   }
-
-  if (raster::nlayers(x) == 1) {
     if (isTRUE(discrete)) {
       # get rasterlabels
       legend_labels <- tryCatch({
@@ -57,69 +61,104 @@ util_plot <- function(x,
         x@data@attributes[[1]][, 2] <- levels
       })
 
-      rasterVis::gplot(x, maxpixel = raster::ncell(x)) +
-        ggplot2::geom_raster(ggplot2::aes(fill = factor(value))) +
+      xyz  <- raster::as.data.frame(x, xy = TRUE)
+
+      ggplot2::ggplot(xyz) +
+        ggplot2::geom_tile(ggplot2::aes(x,y,fill = xyz[,3]))+
         ggplot2::labs(x = xlab,
                       y = ylab)  +
         theme_nlm_discrete(..., legend_labels = legend_labels, ratio = ratio)
 
     } else {
-      rasterVis::gplot(x, maxpixel = raster::ncell(x)) +
-        ggplot2::geom_raster(ggplot2::aes(fill = value)) +
+
+      xyz  <- raster::as.data.frame(x, xy = TRUE)
+
+      ggplot2::ggplot(xyz) +
+        ggplot2::geom_tile(ggplot2::aes(x,y,fill = xyz[,3]))+
         ggplot2::labs(x = xlab,
                       y = ylab) +
         theme_nlm(..., ratio = ratio)
     }
-  } else {
-    rasterVis::levelplot(x)
-  }
 }
 
-
-#' @rdname util_plot
+#' @name show_landscape
 #' @export
-util_plot_grey <- function(x,
-                           xlab = "Easting",
-                           ylab = "Northing",
-                           discrete = FALSE,
-                           ...) {
-  # derive ratio for plot, cells should be a square and axis equal in length
-  if (raster::ncol(x) == raster::nrow(x)) {
-    ratio <- 1
-  } else {
-    ratio <- raster::nrow(x) / raster::ncol(x)
+show_landscape.list <- function(x,
+                                xlab = "Easting",
+                                ylab = "Northing",
+                                discrete = FALSE,
+                                div_scales = FALSE,
+                                n_col = NULL,
+                                n_row = NULL,
+                                ...) {
+
+  if (!div_scales){
+    x_tibble <- tibble::enframe(x, "id", "maps") %>%
+      dplyr::mutate(maps = purrr::map(.$maps, util_raster2tibble)) %>%
+      tidyr::unnest()
+
+    p <- ggplot2::ggplot(x_tibble, ggplot2::aes_string("x", "y")) +
+      ggplot2::coord_fixed() +
+      ggplot2::geom_raster(ggplot2::aes_string(fill = "z")) +
+      ggplot2::facet_wrap(~id, nrow = n_row, ncol = n_col) +
+      ggplot2::scale_x_continuous(expand = c(0, 0), limits = c(0,max(x_tibble$x))) +
+      ggplot2::scale_y_continuous(expand = c(0, 0), limits = c(0,max(x_tibble$y))) +
+      ggplot2::guides(fill = FALSE) +
+      ggplot2::labs(titel = NULL, x = NULL, y = NULL) +
+      theme_facetplot()
   }
 
-  if (raster::nlayers(x) == 1) {
-    if (isTRUE(discrete)) {
-      # discrete case
+  if (div_scales){
 
-      # get rasterlabels
-      legend_labels <- tryCatch({
-        x@data@attributes[[1]][, 2]
-      },
-      error = function(e) {
-        x <- raster::as.factor(x)
-        levels <- raster::unique(x)
-        x@data@attributes[[1]][, 2] <- levels
-      })
+    landscape_plots <- purrr::map(seq_along(x), function(id){
 
-      rasterVis::gplot(x, maxpixel = raster::ncell(x)) +
-        ggplot2::geom_raster(ggplot2::aes(fill = factor(value))) +
-        ggplot2::labs(x = xlab,
-                      y = ylab)  +
-        theme_nlm_grey_discrete(...,
-                                legend_labels = legend_labels,
-                                ratio = ratio)
+      x_tibble <- tibble::enframe(x[id], "id", "maps") %>%
+        dplyr::mutate(maps = purrr::map(.$maps, util_raster2tibble)) %>%
+        tidyr::unnest()
 
-    } else {
-      rasterVis::gplot(x) +
-        ggplot2::geom_raster(ggplot2::aes(fill = value)) +
-        ggplot2::labs(x = xlab,
-                      y = ylab) +
-        theme_nlm_grey(..., ratio = ratio)
-    }
-  } else {
-    rasterVis::levelplot(x)
+      x_tibble$id = names(x[[id]])
+
+      p <- ggplot2::ggplot(x_tibble, ggplot2::aes_string("x", "y")) +
+        ggplot2::coord_fixed() +
+        ggplot2::geom_raster(ggplot2::aes_string(fill = "z")) +
+        ggplot2::facet_wrap(~id, nrow = 1, ncol = 1) +
+        ggplot2::scale_x_continuous(expand = c(0, 0), limits = c(0,max(x_tibble$x))) +
+        ggplot2::scale_y_continuous(expand = c(0, 0), limits = c(0,max(x_tibble$y))) +
+        ggplot2::guides(fill = FALSE) +
+        ggplot2::labs(titel = NULL, x = NULL, y = NULL) +
+        theme_facetplot() +
+        ggplot2::theme(plot.margin = ggplot2::unit(c(0, 0, 0, 0), "cm"))
+
+    })
+
+    p <- cowplot::plot_grid(plotlist = landscape_plots,
+                            nrow = 1,
+                            ncol = 2,
+                            align = "hv"
+    )
+
+    p <- gridExtra::grid.arrange(grobs = landscape_plots, nrow=n_row, ncol=n_col)
   }
+
+  return(p)
+
+}
+
+#' @name show_landscape
+#' @export
+show_landscape.RasterStack <- function(x,
+                                xlab = "Easting",
+                                ylab = "Northing",
+                                discrete = FALSE,
+                                div_scales = FALSE,
+                                n_col = NULL,
+                                n_row = NULL,
+                                ...) {
+
+  maplist <- list()
+  for (i in seq_len(raster::nlayers(x))) {
+    maplist <- append(maplist, list(raster::raster(x, layer = i)))
+  }
+  x <- magrittr::set_names(maplist, names(x))
+  show_landscape.list(x,xlab,ylab, discrete, div_scales,, n_col, n_row, ...)
 }