+++ tests/testthat/test_utilplot.R
@@ -1,18 +1,16 @@
 # nolint start
-context("util_plot")
+context("show_landscape")
 
 x <- rndmap
 
-p <- util_plot(x)
+p <- show_landscape(x)
 
 test_that("basic functionality", {
-  expect_error(util_plot(x), NA)
-  expect_error(util_plot(x, discrete = TRUE), NA)
-  expect_error(util_plot_grey(x), NA)
-  expect_error(util_plot_grey(x, discrete = TRUE), NA)
+  expect_error(show_landscape(x), NA)
+  expect_error(show_landscape(x, discrete = TRUE), NA)
 })
 
-test_that("util_plot behaves like it should", {
+test_that("show_landscape behaves like it should", {
   expect_equal(class(p), c("gg","ggplot"))
 })